Moved traits to top level.
