+++ src/main/scala/org/economicsl/auctions/multiunit/LimitAskOrder.scala
@@ -17,7 +17,7 @@ package org.economicsl.auctions.multiunit
 
 import java.util.UUID
 
-import org.economicsl.auctions.{Price, Quantity, Tradable}
+import org.economicsl.auctions.{Price, Quantity, SinglePricePoint, Tradable}
 
 
 /** An order to sell multiple units of a tradable at a per-unit price greater than or equal to the limit price. */
