+++ src/main/scala/org/economicsl/auctions/multiunit/LimitBidOrder.scala
@@ -17,7 +17,7 @@ package org.economicsl.auctions.multiunit
 
 import java.util.UUID
 
-import org.economicsl.auctions.{Price, Quantity, Tradable}
+import org.economicsl.auctions.{Price, Quantity, SinglePricePoint, Tradable}
 
 
 /** An order to buy multiple units of a tradable at a per-unit price less than or equal to the limit price. */
