+++ src/main/scala/org/economicsl/auctions/multiunit/MarketAskOrder.scala
@@ -17,7 +17,7 @@ package org.economicsl.auctions.multiunit
 
 import java.util.UUID
 
-import org.economicsl.auctions.{Price, Quantity, Tradable}
+import org.economicsl.auctions.{Price, Quantity, SinglePricePoint, Tradable}
 
 
 /** An order to sell a multiple units of a tradable at any positive price. */
